[Refactor] Extract Azure DevOps credential validation (#916)
* refactor: extract Azure DevOps credentials

* chore: remove stale ADO test import

---------
